More optimizated!

+++ src/common/packets/avalibile.rs
@@ -2,6 +2,6 @@ use bytes_kman::prelude::*;
 
 #[derive(Bytes, Clone, Debug)]
 pub struct Avalibile {
-    pub session: u128,
+    pub session: usize,
     pub port: u16,
 }
