+++ src/common/packets/info_request.rs
@@ -4,5 +4,5 @@ use bytes_kman::prelude::*;
 #[derive(Bytes, Clone, Debug)]
 pub struct InfoRequest {
     pub adress: Adress,
-    pub session: u128,
+    pub session: usize,
 }
