+++ src/common/packets/register_response.rs
@@ -3,5 +3,5 @@ use bytes_kman::prelude::*;
 #[derive(Bytes, Clone, Debug)]
 pub struct RegisterResponse {
     pub accepted: bool,
-    pub session: u128,
+    pub session: usize,
 }
