+++ src/common/packets/request.rs
@@ -3,14 +3,14 @@ use bytes_kman::prelude::*;
 
 #[derive(Bytes, Clone, Debug)]
 pub struct Request {
-    pub session: u128,
+    pub session: usize,
     pub to: Adress,
     pub secret: String,
 }
 
 #[derive(Bytes, Clone, Debug)]
 pub struct NewRequest {
-    pub session: u128,
+    pub session: usize,
     pub from: Adress,
     pub secret: String,
 }
