+++ src/common/packets/request_final.rs
@@ -4,14 +4,14 @@ use crate::common::adress::Adress;
 
 #[derive(Bytes, Clone, Debug)]
 pub struct RequestFinal {
-    pub session: u128,
+    pub session: usize,
     pub to: Adress,
     pub accepted: bool,
 }
 
 #[derive(Bytes, Clone, Debug)]
 pub struct NewRequestFinal {
-    pub session: u128,
+    pub session: usize,
     pub from: Adress,
     pub accepted: bool,
 }
